+++ core/amber/src/main/scala/edu/uci/ics/texera/workflow/operators/typecasting/TypeCastingOpDesc.java
@@ -4,7 +4,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyDescription;
 import com.google.common.base.Preconditions;
 import com.kjetland.jackson.jsonSchema.annotations.JsonSchemaTitle;
-import edu.uci.ics.texera.workflow.common.AttributeTypeUtils;
 import edu.uci.ics.texera.workflow.common.metadata.InputPort;
 import edu.uci.ics.texera.workflow.common.metadata.OperatorGroupConstants;
 import edu.uci.ics.texera.workflow.common.metadata.OperatorInfo;
@@ -12,15 +11,11 @@ import edu.uci.ics.texera.workflow.common.metadata.OutputPort;
 import edu.uci.ics.texera.workflow.common.metadata.annotations.AutofillAttributeName;
 import edu.uci.ics.texera.workflow.common.operators.OneToOneOpExecConfig;
 import edu.uci.ics.texera.workflow.common.operators.map.MapOpDesc;
-import edu.uci.ics.texera.workflow.common.tuple.schema.Attribute;
 import edu.uci.ics.texera.workflow.common.tuple.schema.AttributeType;
+import edu.uci.ics.texera.workflow.common.tuple.schema.AttributeTypeUtils;
 import edu.uci.ics.texera.workflow.common.tuple.schema.Schema;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
 import static java.util.Collections.singletonList;
-import static java.util.stream.Collectors.toList;
 import static scala.collection.JavaConverters.asScalaBuffer;
 
 public class TypeCastingOpDesc extends MapOpDesc {
@@ -41,7 +36,7 @@ public class TypeCastingOpDesc extends MapOpDesc {
         if (attribute == null) {
             throw new RuntimeException("TypeCasting: attribute is null");
         }
-        return new OneToOneOpExecConfig(operatorIdentifier(),worker -> new TypeCastingOpExec(this));
+        return new OneToOneOpExecConfig(operatorIdentifier(), worker -> new TypeCastingOpExec(this));
     }
 
     @Override
@@ -57,7 +52,7 @@ public class TypeCastingOpDesc extends MapOpDesc {
     @Override
     public Schema getOutputSchema(Schema[] schemas) {
         Preconditions.checkArgument(schemas.length == 1);
-        if (this.resultType != null){
+        if (this.resultType != null) {
             return AttributeTypeUtils.SchemaCasting(schemas[0], this.attribute, this.resultType);
         }
