+++ core/amber/src/main/scala/edu/uci/ics/texera/workflow/operators/typecasting/TypeCastingOpExec.java
@@ -1,8 +1,8 @@
 package edu.uci.ics.texera.workflow.operators.typecasting;
 
-import edu.uci.ics.texera.workflow.common.AttributeTypeUtils;
 import edu.uci.ics.texera.workflow.common.operators.map.MapOpExec;
 import edu.uci.ics.texera.workflow.common.tuple.Tuple;
+import edu.uci.ics.texera.workflow.common.tuple.schema.AttributeTypeUtils;
 import scala.Function1;
 
 import java.io.Serializable;
