+++ core/amber/src/test/scala/edu/uci/ics/texera/unit/AttributeTypeUtilsTest.scala
@@ -1,6 +1,9 @@
 package edu.uci.ics.texera.unit
 import edu.uci.ics.texera.workflow.common.tuple.schema.AttributeType._
-import edu.uci.ics.texera.workflow.common.AttributeTypeUtils.{inferField, inferSchemaFromRows}
+import edu.uci.ics.texera.workflow.common.tuple.schema.AttributeTypeUtils.{
+  inferField,
+  inferSchemaFromRows
+}
 import org.scalatest.funsuite.AnyFunSuite
 
 class AttributeTypeUtilsTest extends AnyFunSuite {
